Backend/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.0
python-dotenv==1.0.1
google-generativeai==0.8.3
googlemaps==4.10.0
